Fixed PagingAllocator in cooperation with FCFS scheduler

+++ Process.cpp
@@ -2,7 +2,7 @@
 #include "Process.h"
 
 // Constructor implementation.
-Process::Process( int process_id, const std::string &process_name, const std::string &creation_time, 
+Process::Process( size_t process_id, const std::string &process_name, const std::string &creation_time, 
     int core_id, int min_instructions, int max_instructions, size_t min_mem, size_t max_mem )
   : process_id(process_id), process_name(process_name), creation_time(creation_time), 
   core_id(core_id), process_state(READY), requirement_flags{true, 0}
@@ -75,7 +75,7 @@ void Process::setState(ProcessState state)
 }
 
 // Returns the unique process ID (PID).
-int Process::getPID() const
+size_t Process::getPID() const
 {
   return process_id;
 }
